backend/requirements.txt
asgiref==3.5.0
attrs==21.4.0
bleach==4.1.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
distlib==0.3.4
Django==3.2.12
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-spectacular==0.22.0
filelock==3.6.0
identify==2.4.12
idna==3.3
inflection==0.5.1
jsonschema==4.4.0
nodeenv==1.6.0
packaging==21.3
Pillow==9.0.1
platformdirs==2.5.1
pre-commit==2.18.1
PyJWT==2.3.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.0
webencodings==0.5.1
